Hashing
Requirement:
Hashing
This assignment is for practicing Hash Tables. You are required to
create a Spellchecker and an Anagram checker program. (Anagrams are the
words that can be formed by rearranging the letters of the given word.
E.g., “Art” is an anagram of “Tar”.) There is an “official dictionary
(Small-dictionary25K.txt)” uploaded with this assignment. Valid
anagrams can be assumed to be those that are included in this
dictionary.
You will need to create a hash table to store all these words. You may
need to convert them to a key using some conversion. Once you have the
keys- you can use a suitable hash function to create hash keys and
insert all the words in a hash table. (Hint: You may arrange the
letters each words alphabetically and use that as its key. This way
that all anagrams of the words will have the same key and will be in
the same hash-bucket. Then use a suitable hash function to calculate
the hash-key for the words. You may use separate chaining to handle
collisions.)
For spellcheck as well as anagrams of a given word- the process will
need a simple “search” operation. You will need to arrange the given
word alphabetically and then use the hash function to calculate its
hash-key and then do a search. If search is successful then it will
mean that the spelling is correct, else it’s incorrect. In the same
search operation- you can find all of its anagrams which will have the
same hash-key.
In your program- for a given word either you should be able to
spellcheck it and find all the anagrams of the given word.
1. Input file: Give a list of 10 words including some
misspelled/invalid words.
2. Output file: Create an output file (for all the 10 words) that gives
the results as
a. Spellcheck result: as “Correct spelling” or “Incorrect spelling”.
b. Valid anagrams for both- correct as well as incorrect words.
(Anagrams may serve as candidates
for suggested corrections for incorrect spellings.)